chore: moved query hooks into domain level folder
