+++ src/api/entity/entity.mutations.ts
@@ -1,0 +1,2 @@
+export {};
+
