+++ src/api/entity/entity.queries.ts
@@ -1,0 +1,2 @@
+export {};
+
