+++ src/api/queryKeys.ts
@@ -3,6 +3,7 @@ import type { UUID } from "../models/models";
 export const queryKeys = {
   user: ["user"] as const,
   portfolio: ["portfolio"] as const,
+  portfolioUserId: (userId: UUID) => ["portfolio", "userId", userId] as const,
   topics: ["topics"] as const,
   eventById: (eventId: UUID) => ["event", "id", eventId] as const,
   poolsByEventId: (eventId: UUID) =>
