+++ src/api/rank/rank.mutations.ts
@@ -1,0 +1,2 @@
+export {};
+
