+++ src/api/reward/reward.queries.ts
@@ -1,0 +1,2 @@
+export {};
+
