+++ src/api/subtopic/subtopic.mutations.ts
@@ -1,0 +1,2 @@
+export {};
+
