+++ src/api/subtopic/subtopic.queries.ts
@@ -1,0 +1,2 @@
+export {};
+
