+++ src/api/user/user.mutations.ts
@@ -1,0 +1,2 @@
+export {};
+
