+++ src/api/user/user.queries.ts
@@ -1,7 +1,7 @@
 import { useAuth } from "@clerk/react";
 import { useQuery } from "@tanstack/react-query";
 import { queryKeys } from "../queryKeys";
-import { fetchCurrentUser, type CurrentUser } from "../../api/user.api";
+import { fetchCurrentUser, type CurrentUser } from "./user.api";
 
 const LOADING_USER_QUERY_RESULT: CurrentUser = {
   id: "00000000-0000-0000-0000-000000000000",
@@ -19,10 +19,10 @@ export function useUserQuery() {
     queryKey: queryKeys.user,
     queryFn: () => fetchCurrentUser(getToken),
     placeholderData: () => LOADING_USER_QUERY_RESULT,
-    
-    // caching knobs
-    staleTime: 60_000, // 1 min: switching tabs won't refetch immediately
-    gcTime: 10 * 60_000, // 10 min: keep cache around after unmount
+
+    staleTime: 60_000,
+    gcTime: 10 * 60_000,
     refetchOnWindowFocus: false,
   });
 }
+
