+++ src/components/EventCard.tsx
@@ -5,9 +5,9 @@ import useMediaQuery from "@mui/material/useMediaQuery";
 import { useTheme } from "@mui/material/styles";
 import type { Event, Pool } from "../models/models";
 import { useApi } from "../hooks/useApi";
-import { usePoolsQuery } from "../hooks/queries/usePoolsQuery";
+import { usePoolsQuery } from "../api/market/market.queries";
 import { useQueryClient } from "@tanstack/react-query";
-import { queryKeys } from "../hooks/queryKeys";
+import { queryKeys } from "../api/queryKeys";
 import { EventCardLayout } from "./event-card/EventCardLayout";
 import { DesktopEventCardBody } from "./event-card/DesktopEventCardBody";
 import { EventCardFooter } from "./event-card/EventCardFooter";
