+++ src/components/event-card/DesktopEventCardBody.tsx
@@ -3,7 +3,7 @@ import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import { entityTextColor } from "../../utils/entityTextColor";
-import type { PoolItem } from "../../hooks/queries/usePoolsQuery";
+import type { Pool } from "../../models/models";
 
 const scrollableSx = {
   maxHeight: 90,
@@ -19,7 +19,7 @@ export function DesktopEventCardBody({
   isMulti,
   onSelectPool,
 }: {
-  pools: PoolItem[];
+  pools: Pool[];
   totalVolume: number;
   isMulti: boolean;
   onSelectPool: (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement>, poolId: string) => void;
