+++ src/views/LeaderboardView.tsx
@@ -5,7 +5,7 @@ import Divider from "@mui/material/Divider";
 import Skeleton from "@mui/material/Skeleton";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { useLeaderboardQuery } from "../hooks/queries/useLeaderboardQuery";
+import { useLeaderboardQuery } from "../api/rank/rank.queries";
 import type { LeaderboardEntry } from "../models/models";
 
 const LeaderboardLoadingSkeleton = () =>
